Move tree-view-render to components/renders/tree-view folder
