+++ src/components/renders/tree-view/types.ts
@@ -1,4 +1,4 @@
-import { TreeXItemModel } from "../tree-view/tree-view/types";
+import { TreeXItemModel } from "../../tree-view/tree-view/types";
 
 export type TreeXItemModelExtended = {
   parentItem: TreeXItemModel;
