+++ src/components/tree-view/tree-view/readme.md
@@ -72,7 +72,7 @@ Type: `Promise<void>`
 
 ### Used by
 
- - [ch-tree-view-render](../../test)
+ - [ch-tree-view-render](../../renders/tree-view)
 
 ### Graph
 ```mermaid
